+++ CHANGELOG.md
@@ -7,6 +7,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2024-7-28
+### Added
+- The employee actions dropdown
+- The employee request management page
+- Middleware that reestablishes the Session variables if they are signed in by the cookie
+
+### notes
+- attempting to access the page as a non-employee will redirect you to home
+
 ## [0.2.7] - 2024-7-24
 ### Notes
 - Reset the database